Retry EnumDisplayMonitors and default primary monitor
It has been reported in FancyWM/fancywm#235 that the LINQ where we find
the monitor containing the origin (0, 0) can match nothing, but only
when waking up from sleep.

This can happenen if:
 - EnumDisplayMonitors returned no display monitors
 - There is no monitor at the origin (monitor not initialized?)

To handle both cases, wait up to 5 seconds for EnumDisplayMonitors to
return something, and fallback to the first monitor returned.