Do not keep references to IComVirtualDesktop
Keeping COM objects around means our retry on failure strategy is
compromised. If Explorer crashes, these references are invalidated. We
try to re-create the VirtualDesktopInternal and related COM objects, but
not all of the IComVirtualDesktop.

To fix this, we need to pass around HMONITOR + GUID pairs identifying
the virtual desktop, instead of actual COM objects.